+++ seacatauth/middleware.py
@@ -59,11 +59,9 @@ def private_auth_middleware_factory(app):
 		def has_resource_access(tenant: str, resource: str) -> bool:
 			return rbac_svc.has_resource_access(request.Session.Authorization.Authz, tenant, [resource])
 
-		def is_superuser() -> bool:
-			return has_resource_access("*", "authz:superuser")
-
 		request.has_resource_access = has_resource_access
-		request.is_superuser = is_superuser
+		request.is_superuser = rbac_svc.is_superuser(request.Session.Authorization.Authz)
+		request.can_access_all_tenants = rbac_svc.can_access_all_tenants(request.Session.Authorization.Authz)
 
 		if require_authentication is False:
 			return await handler(request)